coding-requirements.txt
ansible==8.0.0
  ansible-core==2.15.0
    cryptography==41.0.1
      cffi==1.15.1
        pycparser==2.21
    Jinja2==3.1.2
      MarkupSafe==2.1.3
    packaging==23.1
    PyYAML==6.0
    resolvelib==1.0.1
ansible-lint==6.17.1
  ansible-compat==4.1.2
    ansible-core==2.15.0
      cryptography==41.0.1
        cffi==1.15.1
          pycparser==2.21
      Jinja2==3.1.2
        MarkupSafe==2.1.3
      packaging==23.1
      PyYAML==6.0
      resolvelib==1.0.1
    jsonschema==4.17.3
      attrs==23.1.0
      pyrsistent==0.19.3
    packaging==23.1
    PyYAML==6.0
    subprocess-tee==0.4.1
  ansible-core==2.15.0
    cryptography==41.0.1
      cffi==1.15.1
        pycparser==2.21
    Jinja2==3.1.2
      MarkupSafe==2.1.3
    packaging==23.1
    PyYAML==6.0
    resolvelib==1.0.1
  black==23.3.0
    click==8.1.3
    mypy-extensions==1.0.0
    packaging==23.1
    pathspec==0.11.1
    platformdirs==3.5.3
    tomli==2.0.1
  filelock==3.12.2
  jsonschema==4.17.3
    attrs==23.1.0
    pyrsistent==0.19.3
  packaging==23.1
  pathspec==0.11.1
  PyYAML==6.0
  requests==2.31.0
    certifi==2023.5.7
    charset-normalizer==3.1.0
    idna==3.4
    urllib3==2.0.3
  rich==13.4.2
    markdown-it-py==3.0.0
      mdurl==0.1.2
    Pygments==2.15.1
  ruamel.yaml==0.17.31
    ruamel.yaml.clib==0.2.7
  subprocess-tee==0.4.1
  wcmatch==8.4.1
    bracex==2.3.post1
  yamllint==1.32.0
    pathspec==0.11.1
    PyYAML==6.0
bandit==1.7.5
  GitPython==3.1.31
    gitdb==4.0.10
      smmap==5.0.0
  PyYAML==6.0
  rich==13.4.2
    markdown-it-py==3.0.0
      mdurl==0.1.2
    Pygments==2.15.1
  stevedore==5.1.0
    pbr==5.11.1
graphviz==0.20.1
pip-autoremove==0.10.0
  pip==23.1.2
  setuptools==59.6.0
pipdeptree==2.9.0
pylint==2.17.4
  astroid==2.15.5
    lazy-object-proxy==1.9.0
    typing_extensions==4.6.3
    wrapt==1.15.0
  dill==0.3.6
  isort==5.12.0
  mccabe==0.7.0
  platformdirs==3.5.3
  tomli==2.0.1
  tomlkit==0.11.8
pywinrm==0.4.3
  requests==2.31.0
    certifi==2023.5.7
    charset-normalizer==3.1.0
    idna==3.4
    urllib3==2.0.3
  requests-ntlm==1.2.0
    cryptography==41.0.1
      cffi==1.15.1
        pycparser==2.21
    pyspnego==0.9.0
      cryptography==41.0.1
        cffi==1.15.1
          pycparser==2.21
    requests==2.31.0
      certifi==2023.5.7
      charset-normalizer==3.1.0
      idna==3.4
      urllib3==2.0.3
  six==1.16.0
  xmltodict==0.13.0
wheel==0.40.0
